# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HelpMotivateMe.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/"
    }
  ],
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/push/subscribe": {
      "post": {
        "tags": [
          "PushNotifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/push/unsubscribe": {
      "delete": {
        "tags": [
          "PushNotifications"
        ],
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/notifications/push/admin/send-to-user": {
      "post": {
        "tags": [
          "PushNotifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/push/admin/send-to-all": {
      "post": {
        "tags": [
          "PushNotifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToAllRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToAllRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SendPushToAllRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushNotificationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/push/admin/users": {
      "get": {
        "tags": [
          "PushNotifications"
        ],
        "parameters": [
          {
            "name": "hasPush",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPushStatus"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPushStatus"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPushStatus"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/waitlist": {
      "post": {
        "tags": [
          "Waitlist"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistSignupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistSignupRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistSignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/waitlist/check": {
      "get": {
        "tags": [
          "Waitlist"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistCheckResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistCheckResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistCheckResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminUserResponse": {
        "required": [
          "id",
          "email",
          "displayName",
          "isActive",
          "membershipTier",
          "role",
          "createdAt",
          "updatedAt",
          "aiCallsCount",
          "aiTotalCostUsd"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "membershipTier": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "aiCallsCount": {
            "type": "integer",
            "format": "int32"
          },
          "aiTotalCostUsd": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "PushNotificationResult": {
        "required": [
          "totalSubscriptions",
          "successCount",
          "failureCount",
          "errors"
        ],
        "type": "object",
        "properties": {
          "totalSubscriptions": {
            "type": "integer",
            "format": "int32"
          },
          "successCount": {
            "type": "integer",
            "format": "int32"
          },
          "failureCount": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PushSubscriptionKeys": {
        "required": [
          "p256dh",
          "auth"
        ],
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          }
        }
      },
      "PushSubscriptionRequest": {
        "required": [
          "endpoint",
          "keys"
        ],
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/PushSubscriptionKeys"
          }
        }
      },
...
      "SendPushToAllRequest": {
        "required": [
          "title",
          "body"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SendPushToUserRequest": {
        "required": [
          "userId",
          "title",
          "body"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UserPushStatus": {
        "required": [
          "userId",
          "email",
          "hasPushEnabled",
          "subscriptionCount",
          "lastPushSentAt"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "hasPushEnabled": {
            "type": "boolean"
          },
          "subscriptionCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastPushSentAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "WaitlistSignupRequest": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "WhitelistCheckResponse": {
        "required": [
          "canSignup"
        ],
        "type": "object",
        "properties": {
          "canSignup": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}